backend_new/requirements.txt
fastapi
uvicorn
chromadb
sentence-transformers
numpy
scikit-learn
pymongo
python-dotenv
openai>=1.86.0
